third_party/WebKit/LayoutTests/http/tests/devtools/sources/sources-panel-extension-names-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
The test verifies that extension names are resolved properly in navigator view.


Running: testAddExecutionContextBeforeFile
top
  FibExtension
    script.js
